rename designSystem -> componentLibrary
